Refactor main.js for test coverage
