+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "create-release",
   "version": "1.0.0",
-  "description": "Generate a release",
+  "description": "Create a release",
   "main": "dist/index.js",
   "scripts": {
     "lint": "eslint 'src/**.js' 'tests/**.js' --fix",
-    "test": "eslint 'src/**.js' 'tests/**.js' --fix && jest",
+    "test": "eslint 'src/**.js' 'tests/**.js' && jest --coverage",
     "build": "ncc build src/main.js",
     "precommit": "npm run build && git add dist/"
   },
@@ -40,7 +40,7 @@
   "jest": {
     "testEnvironment": "node",
     "collectCoverageFrom": [
-      "src/main.js"
+      "src/create-release.js"
     ],
     "coverageThreshold": {
       "global": {
