+++ tests/create-release.test.js
@@ -3,7 +3,7 @@ jest.mock('@actions/github');
 
 const core = require('@actions/core');
 const { GitHub, context } = require('@actions/github');
-const run = require('../src/main.js');
+const run = require('../src/create-release.js');
 
 /* eslint-disable no-undef */
 describe('Create Release', () => {
